+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/mediacodec/MediaCodecRenderer.java
@@ -720,6 +720,7 @@ public abstract class MediaCodecRenderer extends BaseRenderer {
       bypassBatchBuffer.clear();
       bypassSampleBuffer.clear();
       bypassSampleBufferPending = false;
+      oggOpusAudioPacketizer.reset();
     } else {
       flushOrReinitializeCodec();
     }
@@ -2349,7 +2350,7 @@ public abstract class MediaCodecRenderer extends BaseRenderer {
           if (inputFormat != null
               && inputFormat.sampleMimeType != null
               && inputFormat.sampleMimeType.equals(MimeTypes.AUDIO_OPUS)) {
-            oggOpusAudioPacketizer.packetize(bypassSampleBuffer);
+            oggOpusAudioPacketizer.packetize(bypassSampleBuffer, inputFormat.initializationData);
           }
           if (!bypassBatchBuffer.append(bypassSampleBuffer)) {
             bypassSampleBufferPending = true;