+++ RELEASENOTES.md
@@ -75,6 +75,8 @@
         in offload, then no track will be selected.
     *   Disabling gapless support for offload when pre-API level 33 due to
         playback position issue after track transition.
+    *   Prepend Ogg ID Header and Comment Header Pages to bitstream for
+        offloaded Opus playback in accordance with RFC 7845.
     *   Remove parameter `enableOffload` from
         `DefaultRenderersFactory.buildAudioSink` method signature.
     *   Remove method `DefaultAudioSink.Builder.setOffloadMode`.